Visual Studio Projects: Force build to update to the latest windows-build
- windows-build now supports the latest Visual Studio 2022 17.14.3

Fix: #1216

#1216 reports problem with 17.14.2, but MS had already released 17.14.3
as of 5/28/2025, so that version is what is now supported.

This whole array VS2022 problems comes from the fact that Release
builds of certain static libraries are rejected at link time as
incompatible with those libraries produced by any prior Visual Studio
version.  This problem has existed for a very long time (going back to
at least VS2017).  The earlier strategy implemented with simh builds
was to not support the latest Visual Studio tool set until it stopped
changing.  That was sort of OK when Visual Studio versions changed
every 2 years, but since VS2022 was dragging on with frequent changes
support for intermediate versions was added some 3 years back.