SCP: Gather Windows Platform details by alternate means when WMIC is gone
SIMH SHOW VERSION output displays as much potentially useful information
about both the currently running simulator and the host platform it is running
on.  On Windows host systems, some of this information has been gathered
from the output of the WMIC command which is deprecated and is already
being removed from new and future systems.

This change leverages WMIC when it is available, and attempts to gather the
interesting information from elsewhere when it is not.

As reported in #1227