Visual Studio Projects: Fetch windows-build before reporting update needed
The value in _X_REQUIRED_WINDOWS_BUILD normally forces a fetch of the
latest windows_build components that are available.  This value might
not be updated before the local system's Visual Studio tool chain is
updated.  When this Visual Studio is a version that isn't supported by
the locally available windows_build, it is worth checking to see if
windows_build has already been updated to support the newest Visual
Studio version.